inclui SafeVault no projeto, identa e inicia CarregarGrupos na TelaRegistro, identa e clarifica mensagens de erro em RestoreValidateSuite, ativa visibilidade em todas as telas em SystemController, typecastfix em AutenticationController, ienta e verifica leitura de conteúdo para uso na RestoreValidateSuite.RestorePrivateKey na TesteUnitarioProgram (onde há bug atual)

+++ controler/AutenticationControler.java
@@ -31,7 +31,7 @@ public class AutenticationControler {
     public static boolean AuthenticateStep2(String[] possibilidades){
         boolean result = false;
         String nomeUsuario = SystemControler.getNome();
-        String[] senhasPossiveis = new String[Math.pow(2,possibilidades.length)];
+        String[] senhasPossiveis = new String[(int) Math.pow(2,possibilidades.length)];
 
         String[] escolhas = new String[possibilidades.length];
         for(int i = 0; i<possibilidades.length; i++){
